test(csv-parser-price): Migrate integration tests
affects: @commercetools/csv-parser-price

Migrate CLI tests to integration-test directory,
fix apiConfig in cli.js,
add separate host options/flags for apiUrl and authUrl

27
