+++ packages/csv-parser-price/test/helpers/wrong-type-sample.csv
@@ -1,0 +1,3 @@
+variant-sku,customType,customField.foo
+my-price,custom-type,will
+my-price,custom-type,wills
